requirements.txt
torch
transformers
numpy
tqdm
seqeval
datasets
